Use size_t as value type

+++ src/HealthGPS/analysis_module.h
@@ -46,7 +46,7 @@ class AnalysisModule final : public UpdatableModule {
     WeightModel weight_classifier_;
     DoubleAgeGenderTable residual_disability_weight_;
     std::vector<std::string> channels_;
-    std::unordered_map<std::string, int> channel_index_;
+    std::unordered_map<std::string, size_t> channel_index_;
     unsigned int comorbidities_;
     std::string name_{"Analysis"};
     std::vector<core::Identifier> factors_to_calculate_ = {"Gender"_id, "Age"_id};
